api/pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "pratico-admin-api"
version = "0.1.0"
description = "Pratico Admin API (FastAPI)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "SQLAlchemy>=2.0",
  "alembic>=1.13",
  "psycopg[binary]>=3.2",
  "python-multipart>=0.0.9",
  "pydantic>=2.8",
  "python-jose[cryptography]>=3.3",
  "passlib[argon2]>=1.7",
  "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
